+++ http/review.http
@@ -1,8 +1,8 @@
-@reviewId = 31146295-ab5c-4e48-9a53-0647562491c7
+@reviewId = 31146295-ab5c-4e48-9a53-0647562491c1
 @orderId = variable value
 @replyId = variable value
 @storeId = ee334491-78ef-4aaa-886d-fd0ce593d909
-@customerAuthorization =eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwidXNlclJvbGUiOiJDVVNUT01FUiIsImlhdCI6MTc1OTgxMzM0NCwiZXhwIjoxNzU5ODE1MTQ0fQ.AgK3Ya0S5vLc6oneBwP1Vp-R81uq468TRqAAXC1k7g8
+@customerAuthorization = eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwidXNlclJvbGUiOiJDVVNUT01FUiIsImlhdCI6MTc1OTgzMjQyMCwiZXhwIjoxNzU5ODM0MjIwfQ.fshrZH2GnBtFZ7rhMxZJSbNby1pC38IlfB0pRP3RMWY
 @ownerAuthorization = eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwidXNlclJvbGUiOiJPV05FUiIsImlhdCI6MTc1OTgxMzE4NiwiZXhwIjoxNzU5ODE0OTg2fQ.suRnSPONgWnfZ7rpdjqn8W1hFAF5FQijSjJrIxfmKgs
 @managerAuthorization = eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwidXNlclJvbGUiOiJNQU5BR0VSIiwiaWF0IjoxNzU5ODE1MjcxLCJleHAiOjE3NTk4MTcwNzF9.LxhNmkiT2KlJ-LoxPrP_4jX3n7XOWv83IUkyhDnNpws
